Categories Ready, yeah

+++ src/main/java/com/project/demo/rest/category/CategoryController.java
@@ -1,11 +1,20 @@
 package com.project.demo.rest.category;
+import com.project.demo.logic.entity.http.GlobalResponseHandler;
+import com.project.demo.logic.entity.http.Meta;
 import com.project.demo.logic.entity.category.Category;
 import com.project.demo.logic.entity.category.CategoryRepository;
 import com.project.demo.logic.entity.user.User;
+import com.project.demo.logic.entity.user.UserRepository;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
-
+import java.util.Optional;
 import java.util.List;
 @RestController
 @RequestMapping("/category")
@@ -13,11 +22,30 @@ public class CategoryController {
     @Autowired
     private CategoryRepository categoryRepository;
 
+
+    @Autowired
+    private UserRepository userRepository;
     @GetMapping
-    @PreAuthorize("hasAnyRole('ADMIN', 'SUPER_ADMIN', 'USER')")
-    public List<Category> getAllCategories(){
-        return categoryRepository.findAll();
+    @PreAuthorize("isAuthenticated()")
+    public ResponseEntity<?> getAll(
+            @RequestParam(defaultValue = "1") int page,
+            @RequestParam(defaultValue = "10") int size,
+            HttpServletRequest request) {
+
+        Pageable pageable = PageRequest.of(page-1, size);
+        Page<Category> categoriesPage = categoryRepository.findAll(pageable);
+        Meta meta = new Meta(request.getMethod(), request.getRequestURL().toString());
+        meta.setTotalPages(categoriesPage.getTotalPages());
+        meta.setTotalElements(categoriesPage.getTotalElements());
+        meta.setPageNumber(categoriesPage.getNumber() + 1);
+        meta.setPageSize(categoriesPage.getSize());
+
+        return new GlobalResponseHandler().handleResponse("Categories retrieved successfully",
+                categoriesPage.getContent(), HttpStatus.OK, meta);
     }
+    /*public List<Category> getAllCategories(){
+        return categoryRepository.findAll();
+    }*/
 
     @GetMapping("/{id}")
     public Category getCategoryById(@PathVariable Long id) {
@@ -25,29 +53,46 @@ public class CategoryController {
     }
     @PutMapping("/{id}")
     @PreAuthorize("hasAnyRole('SUPER_ADMIN')")
-    public Category updateGame(@PathVariable Long id, @RequestBody Category category) {
-        return categoryRepository.findById(id)
-                .map(existingCategory -> {
-                    existingCategory.setName(category.getName());
-                    existingCategory.setDescription(category.getDescription());
-                    return categoryRepository.save(existingCategory);
-                })
-                .orElseGet(() -> {
-                    category.setId(id);
-                    return categoryRepository.save(category);
-                });
+
+    public ResponseEntity<?> updateCategory(@PathVariable Long id, @RequestBody Category category, HttpServletRequest request) {
+        Optional<Category> foundCategory = categoryRepository.findById(id);
+        if(foundCategory.isPresent()) {
+            category.setId(foundCategory.get().getId());
+            categoryRepository.save(category);
+            return new GlobalResponseHandler().handleResponse("Category updated successfully",
+                    category, HttpStatus.OK, request);
+        } else {
+            return new GlobalResponseHandler().handleResponse("Order id " + id + " not found"  ,
+                    HttpStatus.NOT_FOUND, request);
+        }
     }
 
     @PostMapping
     @PreAuthorize("hasAnyRole('SUPER_ADMIN')")
-    public Category addProduct(@RequestBody Category category) {
-        return  categoryRepository.save(category);
+    public ResponseEntity<?> addCategory(@RequestBody Category category, HttpServletRequest request){
+        Optional<Category> foundCategory = categoryRepository.findById(category.getId());
+        if(foundCategory.isPresent()) {
+            Category savedCategory = categoryRepository.save(category);
+            return new GlobalResponseHandler().handleResponse("Category created successfully",
+                    savedCategory, HttpStatus.CREATED, request);
+        }else return new GlobalResponseHandler().handleResponse("Category id " + category.getId() + " not found"  ,
+                HttpStatus.NOT_FOUND, request);
     }
 
     @PreAuthorize("hasAnyRole('SUPER_ADMIN')")
     @DeleteMapping("/{id}")
-    public void deleteCategory (@PathVariable Long id) {
-        categoryRepository.deleteById(id);
+    public ResponseEntity<?> deleteCategory(@PathVariable Long id, HttpServletRequest request){
+        Optional<Category> foundCategory = categoryRepository.findById(id);
+        if(foundCategory.isPresent()) {
+            categoryRepository.deleteById(id);
+            return new GlobalResponseHandler().handleResponse("Category deleted successfully",
+                    HttpStatus.OK, request);
+        }else return new GlobalResponseHandler().handleResponse("Category id " + id + " not found"  ,
+                HttpStatus.NOT_FOUND, request);
     }
+
+    /*public void deleteCategory (@PathVariable Long id) {
+        categoryRepository.deleteById(id);
+    }*/
 
 }
